+++ VerdeterCommand_test.go
@@ -21,3 +21,21 @@ func TestNewVerdeterCommand(t *testing.T) {
 	)
 	assert.NotNil(t, verdeterCmd, "func NewVerdeterCommand should not return something nil")
 }
+
+func TestNewVerdeterCommandLookup(t *testing.T) {
+	verdeterCmd := verdeter.NewVerdeterCommand(
+		"test",
+		"test short description",
+		"test long description",
+		func(verdeterCmd *verdeter.VerdeterCommand, args []string) error {
+			fmt.Println("hello mom")
+			return nil
+		},
+	)
+
+	assert.Nil(t, verdeterCmd.Lookup("some.config"))
+	verdeterCmd.GKey("some.config", verdeter.IsInt, "", "some.config is an simple config key for a unit test")
+	if assert.NotNil(t, verdeterCmd.Lookup("some.config")) {
+		assert.Equal(t, "some.config", verdeterCmd.Lookup("some.config").Name)
+	}
+}